상세, 목록, 검색 구현
