+++ semiproject/src/main/java/com/kh/semi/controller/MainController.java
@@ -1,13 +1,13 @@
 package com.kh.semi.controller;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
 public class MainController {
 
-	@GetMapping("/main")
-	public String main() {
+	@RequestMapping("/")
+	public String main() {  
 		return "/WEB-INF/views/main.jsp";
 	}
-	
 }
