+++ semiproject/src/main/resources/static/css/review.css
@@ -1,0 +1,93 @@
+.input-file-button{
+            text-align: center;
+            cursor: pointer;
+            border-radius: 50px;
+            font-size: 13px;
+            padding: 0.75em;
+            padding-left: 1.5em;
+            padding-right: 1.5em;
+            border-width: 2px;
+            border-style: solid;
+            background-color: #776BFF;
+            border-color: #776BFF;
+            color: white;
+        }
+
+        #chooseFile{
+            visibility: hidden;
+        }
+        
+      	/*
+      		리뷰 별점
+      	*/
+
+        .reviewStar .starR{
+            text-decoration: none;
+            color: #776BFF;
+        }
+        .reviewStar .starR.on{
+            color: #776BFF;
+        }
+        
+        .review-star .starR{
+            text-decoration: none;
+            color: #776BFF;
+        }
+        .review-star .starR.on{
+            color: #776BFF;
+        }
+        
+        /*
+        	내 리뷰 목록 테이블 
+        */
+        
+        .fa-star{
+        	color: #776BFF;
+        }
+        
+        .reviewContent a{
+        	text-decoration: none;
+        }
+        
+        .reviewTime{
+        	font-size:0.7em;
+        	color: gray;
+        }
+        
+        /*
+        	리뷰창 
+        */
+        
+        .review-item{
+        	font-size:15px;
+        	padding:1em;
+        }
+        .review-item .memberId{
+        	font-size: 1.15em
+        }
+        .review-item .reviewContent{
+        	font-size:1em;
+        	min-height:2em;
+        }
+        .review-item .reviewTime{
+        	font-size:0.7em;
+        	color: gray;
+        }
+         .review-item .changeText .text{
+        	font-size: 12px;
+        	color: gray;
+        	cursor: pointer;
+        }
+         .review-item .changeText .text:hover{
+        	color: #776BFF;
+        	cursor: pointer;
+        }
+        
+        /*
+        	리뷰 좋아요
+        */
+        
+        .review-item .fa-heart {
+        	color:red;
+        	cursor: pointer;
+        }
