+++ semiproject/src/main/webapp/WEB-INF/views/main.jsp
@@ -3,8 +3,10 @@
     
 <jsp:include page="/WEB-INF/views/template/header.jsp"></jsp:include>
 		
-      <div class="container-1000">
-          Elit pellentesque habitant morbi tristique senectus et netus et malesuada. Iaculis eu non diam phasellus vestibulum. Sed adipiscing diam donec adipiscing tristique risus nec feugiat in. Auctor eu augue ut lectus arcu bibendum. Vel quam elementum pulvinar etiam non quam. Elementum pulvinar etiam non quam lacus suspendisse faucibus. Fringilla ut morbi tincidunt augue interdum velit. Cursus in hac habitasse platea dictumst. Pharetra vel turpis nunc eget lorem dolor. Quam lacus suspendisse faucibus interdum posuere lorem ipsum dolor. Nunc sed velit dignissim sodales ut. Tortor at auctor urna nunc id cursus metus. Faucibus turpis in eu mi. Diam in arcu cursus euismod quis viverra.
-      </div>
+            <div class="container-1000">
+                <div class="row pt-10">
+                    Elit pellentesque habitant morbi tristique senectus et netus et malesuada. Iaculis eu non diam phasellus vestibulum. Sed adipiscing diam donec adipiscing tristique risus nec feugiat in. Auctor eu augue ut lectus arcu bibendum. Vel quam elementum pulvinar etiam non quam. Elementum pulvinar etiam non quam lacus suspendisse faucibus. Fringilla ut morbi tincidunt augue interdum velit. Cursus in hac habitasse platea dictumst. Pharetra vel turpis nunc eget lorem dolor. Quam lacus suspendisse faucibus interdum posuere lorem ipsum dolor. Nunc sed velit dignissim sodales ut. Tortor at auctor urna nunc id cursus metus. Faucibus turpis in eu mi. Diam in arcu cursus euismod quis viverra.
+                </div>
+            </div>
 		
 <jsp:include page="/WEB-INF/views/template/footer.jsp"></jsp:include>
